python-services/ml-service/requirements.txt
# 现有依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# 数据处理
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# ML/AI
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.0

# LLM 集成
openai==1.6.0
anthropic==0.8.0
langchain==0.1.0
langchain-community==0.0.10

# 向量数据库
pinecone-client==3.0.0
weaviate-client==3.26.0
chromadb==0.4.18

# 数据质量
great-expectations==0.18.5